backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
